TBR: BR:davemoore@chromium.org
A few days ago, tedbo landed a change that got rid of a step in
customize_rootfs.sh that put the chronos user into a bunch of groups,
including "video" and one called "admin".  He added back in the bit
that put it into the "video" group, which gave chronos permission to
use hardware accelerated video again.  We don't need the "admin" group,
though, so I'm just getting rid of it here.  This refrence was making
this chown fail, which was making /home/chronos continue to be owned by
root.  Thus, we couldn't write into /home/chronos, which broke a lot of stuff.