session_manager will crash chrome if it hasn't exited on time
Back when the session_manager used to handle poking chrome to shut it down,
it'd SIGTERM it, wait a bit, and then SIGKILL.  THat's what Linux usually does.
In our case, if something's making chrome take a while to shut down, we want
to find out.  So, we'd like to generate a crash dump.  To that end, once we
go back to using the session manager to TERM chrome when the user wants to
sign out this change will make the browser get SIGABRT'd if it's taking too
long, generating and uploading a crash dump.

BUG=None
TEST=unit tests

Change-Id: I4391710499d8c6a7d03108239c57b1bf2242d1e3

Review URL: http://codereview.chromium.org/3005056