[login_manager] Add async persisting of Owner key, alter API to allow NULL inside sigs
I was using gchar* as the type for key and signature blobs in my API.  They can have 0x00 in the middle, though, so that doesn't work.  Took a page out of fes' book and decided to use a GArray instead.  Safer that way.

Also made it so that chronos can call my API.  Duh :-)

BUG=4486
TEST=unit tests

Change-Id: I3c85893f0b72238b3a93b86b2c5d1fd8376d4935

Review URL: http://codereview.chromium.org/3165044