[login_manager] Upon ownership, create base policy to whitelist and register owner.
When ownership of a device is taken, we need to create and sign a
policy that whitelists the owner, and registers his username.  We
already do this in the current (deprecate) signed key/value pair
scheme; this change brings the new policy-backed ownership
implementation in line.

BUG=13842
TEST=unit tests
STATUS=Fixed

Change-Id: I4e9167cfb193e348e6c0b511ed7b7bd55d5ec2c1

R=gauravsh@chromium.org

Review URL: http://codereview.chromium.org/6813073